fix(refactor): Shift data file to correct location
