+++ data/LeeKuanYew.txt
@@ -1,0 +1,1 @@
+T|0|Meng Shuo
